+++ app/lib/shared/versions.dart
@@ -24,10 +24,10 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const _acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2026.03.27',
+  '2026.04.07',
   // Fallback runtime versions.
+  '2026.03.27',
   '2026.03.26',
-  '2026.03.17',
 ];
 
 /// Sets the current runtime versions.
@@ -62,8 +62,8 @@ bool shouldGCVersion(String version) =>
 
 // keep in-sync with SDK version in .github/workflows/all-test.yml and Dockerfile.app
 final String runtimeSdkVersion = '3.11.0';
-final String toolStableDartSdkVersion = '3.11.2';
-final String toolStableFlutterSdkVersion = '3.41.4';
+final String toolStableDartSdkVersion = '3.11.4';
+final String toolStableFlutterSdkVersion = '3.41.6';
 
 final semanticToolStableDartSdkVersion = Version.parse(
   toolStableDartSdkVersion,